requirements.txt
streamlit>=1.40,<2
pandas
psycopg[binary]
python-dotenv
pygments
